Add API support for downloading different years of data
This commit does not actually add multiple vintages of data, but it
updates the API of both the geotype classes and the CLI to accept a year
argument. This argument is validated and used to construct source
URLs and filenames.

Setting up actual support for multiple vintages of data will require a
bit of research about which vintage are supported for the different
geotypes and how URLs and filenames may vary over that time period.

Addresses #8
(https://github.com/datadesk/census-map-downloader/issues/8)