Support additional years for counties
Support years 2011-2020 for counties and 2014-2019 for cartographic
counties. These years are based on the somewhat arbitrary choice of
supporting all the years that follow the same URL pattern and field
list as the original supported year, 2018.

This likely does not fully support all the vintages of all the geotypes
that users need, but provides a reference implementation for what
it looks like to add support for additional years.

I updated the README with some suggestions for how to support additional
years for more complex cases.

Addresses #8
(https://github.com/datadesk/census-map-downloader/issues/8)